requirements.txt
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.6
cssselect2==0.4.1
defusedxml==0.7.1
ffmpeg-python==0.2.0
filelock==3.0.12
future==0.18.2
pathlib==1.0.1
pdf2image==1.16.0
Pillow==8.3.2
pycparser==2.20
pyexifinfo==0.4.0
PyPDF2==1.26.0
python-magic==0.4.24
tinycss2==1.1.0
Wand==0.6.7
webencodings==0.5.1
xvfbwrapper==0.2.9
pika==1.2.0
pika-stubs==0.1.3
walrus==0.9.3
